Refactors GPT CLI architecture for better modularity and clarity
- Splits monolithic entry-point script into dedicated modules:
  - `main.py`: argument parsing, REPL loop, command routing
  - `chat.py`: OpenAI communication, memory handling, prompt assembly
  - `logging.py`: prompt/response log file management

- Improves separation of concerns, making the codebase easier to navigate and maintain.

- Adds descriptive docstrings to all functions for clarity and documentation purposes.

- Preserves existing features such as streaming, memory summarization, and REPL support, while enabling future extensibility.

- No changes to CLI behavior or arguments — fully backward-compatible.
