+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "blackcortex-gpt-cli"
-version = "1.2.1"
+version = "1.2.2"
 description = "Terminal-based GPT CLI with memory, markdown, and stream support"
 keywords = ["gpt", "cli", "terminal", "openai"]
 readme = "README.md"
@@ -41,7 +41,7 @@ dev = [
 
 # CLI entry point
 [project.scripts]
-gpt = "blackcortex_cli.gpt:main"
+gpt = "blackcortex_cli.main:main"
 
 # Build system configuration
 [build-system]
